// SistemaVentasBatia/ClientApp/src/app/models/agregarusuario.ts
export interface AgregarUsuario {
    idPersonal: number;
    autorizaVentas: number;
    estatusVentas: number;
    cotizadorVentas: number;
    revisaVentas: number;
    nombres: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    puesto: string;
    telefono: string;
    telefonoExtension: string;
    telefonoMovil: string;
    email: string;
    firma: string;
    usuario: string;
    password: string;
}